Add .DELETE_ON_ERROR
Since we're now using curl to download stock firmware files, we have
to make sure that an error during download will not result in a partial
file being left over. make will automatically delete files when it
receives a signal (Ctrl-C for instance), but it will not when a rule
returns a non-zero status. In order to that we need .DELETE_ON_ERROR.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>