Use new repository for TF-A (trusted-firmware-a) and update TF-A code
Rename submodule and path "arm-trusted-firmware" to "trusted-firmware-a"
(the new official name since quite some time now), and use the official
repository at trustedfirmware.org instead of GitHub. The commit used in
the submodule is updated to the one that is currently being used in
https://github.com/OP-TEE/manifest.

Aliases for Makefile targets "arm-tf" and "arm-tf-clean" are added:
"tfa" and "tfa-clean". No further cleanup is done for simplicity
(variable names...).

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>